moved legacy counter pipelines
